Ent_chatbot/requirements.txt
fastapi
uvicorn
pydantic
boto3
python-dotenv
pyyaml